requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
motor==3.3.2
aiohttp==3.9.1
Pillow==10.1.0
python-dotenv==1.0.0
structlog==23.2.0
asyncio-throttle==1.0.2

